Grid Iteration build 2,0
Bigger tiles
more specific customization
show connections and combos for testing/debugging purposes
seperated combo, connection and grid combo manager into seperate scritps
