+++ Puzzle Game/Assets/Scripts/Grid/GridManager.cs
@@ -10,19 +10,9 @@ public enum ColorEnum {
 
 
 public class GridManager : MonoBehaviour
-{
-
-
-    //BUG, if there is an active line, and then you chnage grid, will create a null reference error  
-    //Ignore diagonal connections  TESTING
-    //TODO automatic drag with mouse, TESTING  
-    //TODO 2 valid connections of the same color type causes a bug, fix should just be clear connectedtiles on valid connection
+{ 
     //TODO Dynamic Dropdown system -> maybe try manipulating grid based on x and y pos, (Seperate Script, TileMovement), "invisible" instantiated row up top.
     //^ algorhthm for droppinng tiles based on connection. Refactor gridManager? 
-    
-    //TODO consider using sets to deal with duplicates in connected tiles script , NOT NEEDED 
-
-    //Seperate Class to measure combos and types of connections 
     //TODO random null reference sometimes ontriggerenter, cant remake bug on command
     [SerializeField]
     private GridComboManager gridComboManager;
